Add tratamento de erro

+++ src/bankAccount/bankAccount.service.ts
@@ -1,8 +1,10 @@
-import { BadRequestException, Inject } from '@nestjs/common';
+import { Inject } from '@nestjs/common';
 import { CreateBankAccountDto } from './dto/create-bank-account.dto';
 import { UpdateBankAccountDto } from './dto/update-bank-account';
 import { BankAccountRepository } from './repositories/bankAccount.repository';
 import { User } from 'src/user/entities/user.entity';
+import { REGISTRATION_TYPE } from 'src/group/enums/registration_type.enum';
+import { CustomError } from 'src/common/error/custom.error';
 
 export class BankAccountService {
   constructor(
@@ -16,19 +18,24 @@ export class BankAccountService {
   ) {
     try {
       const bankAccount = {
-        user_id: { id: String(loggedUserId) } as User,
+        user: { id: String(loggedUserId) } as User,
         ...createBankAccountDto,
-      };
-
-      return this.bankAccountRepository.create(bankAccount);
+      } as CreateBankAccountDto;
+      const result = await this.bankAccountRepository.create(bankAccount);
+      console.log(result);
+      return { success: true, status: 201, data: result };
     } catch (error) {
-      console.log(error);
-      throw new BadRequestException('Erro ao criar conta bancária!');
+      throw new CustomError(
+        [error.message || 'Erro ao incluir conta bancaria'],
+        400,
+        'NEW_BANK_ACCOUNT',
+      );
     }
   }
 
-  async findAll() {
-    return this.bankAccountRepository.findAll();
+  async findAll(loggedUserId: string) {
+    const type = REGISTRATION_TYPE.BANK_ACCOUNT;
+    return this.bankAccountRepository.findAll(loggedUserId, type);
   }
 
   async findById(id: string) {
@@ -39,12 +46,43 @@ export class BankAccountService {
     return bankAccount;
   }
 
-  async update(id: string, bankAccount: UpdateBankAccountDto) {
-    return this.bankAccountRepository.update(id, bankAccount);
+  async update(
+    id: string,
+    bankAccountDto: UpdateBankAccountDto,
+    loggedUserId: string,
+  ) {
+    try {
+      const result = await this.validateLoggedInUser(id, loggedUserId);
+      if (result.inError) {
+        return result.data.message;
+      }
+      return this.bankAccountRepository.update(id, bankAccountDto);
+    } catch (error) {
+      console.error(error);
+      return error;
+    }
   }
 
-  async remove(id: string) {
-    await this.findById(id);
+  async remove(id: string, loggedUserId: string) {
+    const result = await this.validateLoggedInUser(id, loggedUserId);
+    if (result.inError) {
+      return result.data.message;
+    }
     return this.bankAccountRepository.delete(id);
+  }
+
+  async validateLoggedInUser(idBankAccount: string, loggedUserId: string) {
+    const bankAccount =
+      await this.bankAccountRepository.findById(idBankAccount);
+
+    if (bankAccount.user.id != loggedUserId) {
+      return {
+        inError: true,
+        data: {
+          message: 'Usuário não tem permissão para alterar este registro',
+          status: 400,
+        },
+      };
+    }
   }
 }
